requirements.txt
time
Pillow
numpy
selenium